requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
pyyaml>=6.0
protobuf>=4.25.0
grpcio-tools>=1.60.0

